+++ R/zzz.R
@@ -1,8 +1,19 @@
 .onAttach <- function(libname, pkgname) {
   # display the warning message to all users
-
+  
+  # this startup message can be suppressed from version 0.2.4
   packageStartupMessage(cli::rule(left = "WARNING!", line = 2))
-  startup()
+  packageStartupMessage(
+    "The method provided by this package has its limitations ",
+    "and anyone must use them cautiously and responsibly. ",
+    "You should also agree that you only intend to use the method ",
+    "for academic research purpose and not for commercial use. ",
+    "You would also agree NOT to discriminate anyone based on ",
+    "race or color or any characteristic, ",
+    "with the information provided by this package. ",
+    "Please refer to the documentation for details: ",
+    "https://fangzhou-xie.github.io/rethnicity/index.html"
+  )
   packageStartupMessage(cli::rule(line = 2))
   
   invisible()
@@ -10,16 +21,4 @@
 
 .onUnload <- function(libpath) {
   library.dynam.unload("rethnicity", libpath)
-}
-
-startup <- function() {
-  # cli::cli_par()
-  # cli::rule(left = "WARNING!", line = 2)
-  cli::cli_text("The method provided by this package has its limitations and anyone must use them cautiously
-and responsibly. You should also agree that you only intend to use the method for academic
-research purpose and not for commercial use. You would also agree NOT to discriminate anyone
-based on race or color or any characteristic, with the information provided by this package.
-Please refer to the documentation for details: https://fangzhou-xie.github.io/rethnicity/index.html")
-  # cli::rule(line = 2)
-  # cli::cli_end()
 }