+++ NEWS.md
@@ -1,8 +1,21 @@
+# rethnicity (0.2.4)
+
+## fix CRAN notes:
+
+* startup message can be suppressed
+
+## other minor changes:
+
+* update README: it is recommended to use `pak` for installation
+
+
 # rethnicity (0.2.3)
+
 * update dependency on RcppThread >= 2.1.3 to fix the compilation error
 * update metadata of package
 
 # rethnicity (0.2.2)
+
 * add `predict_lastname()` and `predict_fullname()` functions as lower-level interfaces
 * they also allow for customized models
 * add another vignette on how to run inference on the customized models
@@ -11,13 +24,16 @@
 * remove docs/ folder, since we are using github actions to automate the process
 
 # rethnicity (0.1.2)
+
 * fix the WARN message from CRAN checks on the vignette
 
 # rethnicity (0.1.1)
+
 * fixed the include path problem and suppressing warning problem
 * add `cran-comments.md` into `.Rbuildignore`
 * modify the DESCRIPTION file as requested
 
 # rethnicity (0.1.0)
+
 * first submission of the package
 * provide `predict_ethnicity()` function to predict the ethnicity from names